Update README.md
Predicting Bank Loan Elibility
G. MallikarjunaRao
SK. Mohammed Asif
A. Sairam Naik
V. Rakesh
                                                           Introduction

For financial institutions, determining whether a borrower will be eligible for a bank
loan is a crucial step in determining their creditworthiness. In the past, banks have evaluated
loan applications using manual processes, which can be labor-intensive and prone to
mistakes.Large datasets can be analyzed by machine learning to find patterns and trends that
can be used to forecast the likelihood of loan approval, which has the potential to automate and
enhance the loan approval process.The goal of this project is to create a predictive model based
on machine learning algorithms that can reliably forecast loan eligibility based on a variety of
variables, including income, credit score, employment history, and debt-to-income ratio. Banks
and other financial institutions can use the model to reduce the risk of loan default by deciding
whether to approve loan applications.
Like many other business ventures, the banking sector is increasingly looking to take
advantage of the opportunities presented by modern technologies to improve their processes,
productivity and reduce costs. According to [1], the predictive analytics feature of Machine
learning was the most utilized feature for applications in the banking sector worldwide in 2020.
The success or failure of most lending platforms largely depends on their ability to
evaluate credit risk [2].The loan approval process is a challenging task for any financial
institution. Before giving credit loans to borrowers, the bank decides whether the borrower is
bad (defaulter) or good (non-defaulter).
This paper focused on developing Machine Learning (ML) models to predict loan
eligibility, which is vital in accelerating the decision-making process and determining if an
applicant gets a loan or not.
Our objectives in this study include; (1) Clean and Pre-process the data for modeling,
(2) Perform Exploratory Data Analysis (EDA) on the dataset, (3) Build various ML models
to predict loan eligibility, and (4) Evaluate and Compare the different Models built.
The project involves analyzing a dataset of loan applicants and their financial
information, preprocessing and cleaning the data, and selecting relevant features for the
predictive model. Various machine learning algorithms such as logistic regression, decision
trees, and random forests can be employed to create the model. The model's performance can
be evaluated using metrics such as accuracy, precision, recall,
Overall, the development of a predictive model for bank loan eligibility using machine
learning can lead to more efficient and accurate loan approval processes, reduce the risk of
defaults, and help financial institutions to better serve their customers.